test: Remove test subdir in favor of inline mod (#809)
