+++ miden-lib/src/lib.rs
@@ -18,9 +18,6 @@ pub mod accounts;
 pub mod notes;
 pub mod transaction;
 
-#[cfg(all(test, feature = "std"))]
-mod tests;
-
 // RE-EXPORTS
 // ================================================================================================
 
@@ -58,5 +55,29 @@ impl Library for MidenLib {
 
     fn dependencies(&self) -> &[LibraryNamespace] {
         self.contents.dependencies()
+    }
+}
+
+// TESTS
+// ================================================================================================
+
+// NOTE: Most kernel-related tests can be found under /miden-tx/kernel_tests
+#[cfg(all(test, feature = "std"))]
+mod tests {
+    use miden_objects::assembly::Library;
+
+    #[test]
+    fn test_compile() {
+        let path = "miden::kernels::tx::memory::get_input_note_ptr";
+        let miden = super::MidenLib::default();
+        let exists = miden.modules().any(|module| {
+            module
+                .ast
+                .procs()
+                .iter()
+                .any(|proc| module.path.append(&proc.name).unwrap().as_str() == path)
+        });
+
+        assert!(exists);
     }
 }
